Add tests for LCOW shared scratch space work
First test validates that after launching a pod and one container there is only a single
scsi disk available in the guest.

Second test launches two workload containers both sharing the pod sandbox containers scratch space. A
file is written in the first workload container and the available size left on the rootfs is then checked in
both workload containers. The success case is if both containers show the same available size left.

Signed-off-by: Daniel Canter <dcanter@microsoft.com>